+++ src/app/App.tsx
@@ -1,7 +1,15 @@
-import HomePage from '~/components/HomePage';
+import { BrowserRouter, Route, Routes } from 'react-router';
+
+import AppLayout from '~/layouts/app-layout';
 
 function App() {
-    return <HomePage></HomePage>;
+    return (
+        <BrowserRouter>
+            <Routes>
+                <Route path='*' element={<AppLayout />} />
+            </Routes>
+        </BrowserRouter>
+    );
 }
 
 export default App;
