+++ src/components/shared-components/Button/index.ts
@@ -1,0 +1,3 @@
+export * from './Bookmark';
+export * from './Outlined';
+export * from './Rounded';
