+++ src/components/shared-components/StatItem/index.tsx
