+++ src/components/shared-components/Text/index.ts
@@ -1,0 +1,2 @@
+export * from './Regular';
+export * from './Title';
