+++ src/components/shared-components/index.ts
@@ -1,0 +1,8 @@
+export * from './BookmarksSection';
+export * from './Button';
+export * from './ButtonCategory';
+export * from './ButtonViewMore';
+export * from './RecieptCard';
+export * from './Slider';
+export * from './StatItem';
+export * from './Text';
