--- src/components/shared/SectionSubTitle/__index.bak
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import SectionTitle, { SectionTitleProps } from '../Text/Title';
-
-interface SubTitleProps extends SectionTitleProps {
-    title: string;
-    fontSize?: string;
-    fontWeight?: number;
-}
-
-const SubTitle: React.FC<SubTitleProps> = ({ title, fontWeight, ...restProps }) => (
-    <SectionTitle title={title} fontWeight={fontWeight} {...restProps} />
-);
-
-export default SectionSubTitle;
