+++ src/constants/routes.ts
@@ -1,0 +1,27 @@
+export const ROUTES_NAV = {
+    salates: '/salates',
+    zakuski: '/zakuski',
+    pervyeBlyuda: '/pervye-blyuda',
+    vtoryeBlyuda: '/vtorye-blyuda',
+    desertyVypechka: '/deserty-vypechka',
+    blyudaNaGrile: '/blyuda-na-grile',
+    veganskayaKuhnya: {
+        root: '/veganskaya-kuhnya',
+        zakuski: '/veganskaya-kuhnya/zakuski',
+        pervyeBlyuda: '/veganskaya-kuhnya/pervye-blyuda',
+        vtoryeBlyuda: '/veganskaya-kuhnya/vtorye-blyuda',
+        garniry: '/veganskaya-kuhnya/garniry',
+        deserty: '/veganskaya-kuhnya/deserty',
+        vypechka: '/veganskaya-kuhnya/vypechka',
+        syroedcheskieBlyuda: '/veganskaya-kuhnya/syroedcheskie-blyuda',
+        napitki: '/veganskaya-kuhnya/napitki',
+        zagotovki: '/veganskaya-kuhnya/zagotovki',
+    },
+    detskieBlyuda: '/detskie-blyuda',
+    lechebnoePitanie: '/lechebnoe-pitanie',
+    nacionalnyeBlyuda: '/nacionalnye',
+    sousy: '/sousy',
+    napitki: '/napitki',
+    zagotovki: '/zagotovki',
+    root: '*',
+};
