+++ src/views/zakuski/index.tsx
@@ -1,0 +1,6 @@
+import { Box } from '@chakra-ui/react';
+import React from 'react';
+
+const Zakuski: React.FC = () => <Box fontSize={150}>Тут будут закуски</Box>;
+
+export default Zakuski;
